[BACK-145] Add Server API to look for users by IDs
Add a query parameter to the `/users` API endpoint,
where the parameter is `id` and the value is a string
array of Tidepool User IDs.
For example, `/users?id=abc,def,ghi`
This will allow another server to look for a set of users,
rather than needing to call `/userInfo` for each user it
wants to look up.