[BACK-1133] Use Collation to perform case-insensitive queries
Create indexes using collation, and change queries to use collation
in order to do case-insensitive searches whilst being able to use
indexes. The previous method of using a RegEx does _not_
take advantage of indexes.

[BACK-338] Also add a TTL index to automatically remove expired tokens
from the `user.tokesn` collection.
TTL indexes require fields to be of BSON's `Date` type,
so type changes had to be made accordingly.

The only other place this struct is used is in a migration in
`platform/migrations`

* Fixes [BACK-1133]
* Fixes [BACK-338]