upgrade golang.org/x/net package to fix CVE-2024-45338
My investigation indicates that neither our code, nor our dependencies
use the functions in question, but at the same time, the impact of
upgrading them is minimal, so it feels like the less risky path to
just upgrade and not have to worry.

Intentionally not bumping go-common at this time. The version of
go-common used by shoreline is so old that it does not have the
vulnerability, and upgrading it would require additional shoreline
modifications.

BACK-3351